move: 공용컴퍼넌트 폴더 생성 후 해당 파일 이동
폴더가 존재하지 않는 공용컴포넌트의 경우, 공용컴퍼넌트 폴더 생성 후 해당 파일 이동
