+++ haul-fe/src/components/CarInfoBox/CarInfoBox.jsx
@@ -1,9 +1,9 @@
 import styled from "styled-components";
-import theme from "../styles/theme/Theme.jsx";
-import Porter from "../assets/pngs/porter.png";
-import Typography from "./Typhography/Typhography.jsx";
-import Margin from "./Margin/Margin.jsx";
-import UnderBar from "./UnderBar/UnderBar.jsx";
+import theme from "../../styles/theme/Theme.jsx";
+import Porter from "../../assets/pngs/porter.png";
+import Typography from "../Typhography/Typhography.jsx";
+import Margin from "../Margin/Margin.jsx";
+import UnderBar from "../UnderBar/UnderBar.jsx";
 
 const CarTypeImage = {
   포터2: Porter,
